:bug: Fix interrupts hanging processor (#68)
The i2c demo does not work for lpc40 currently and this was due to a bug
with the `hal::cortex_m::is_the_same_vector_buffer()` API. Because the
offset was applied incorrectly, each invocation of this API always
returned false, resulting in the interrupt vector table being reset
each time its called. Meaning any peripheral that initializes the
interrupt vector table, will wipe out all of the entires saved prior.
The interrupts are still enabled, so when they are triggered, the
interrupt invokes default_interrupt_handler which does nothing but spin,
hanging the processor.