:bug: Use new `hal::stm32f1::can_peripheral_manager` ctor in demos
In order to prevent applications unassociated from the can application
to continue to work as usual, a try/catch is put around the construction
of the `hal::stm32f1::can_peripheral_manager` object. If construction
fails, then none of the can interface optionals will be set and the
program will run without them. If the can application is executed and
the optionals are nullopt, then access will throw and will terminate
the application which is the intended behavior.