:bug: Add timeout ctor to `hal::stm32f1::can_peripheral_manager`
`hal::stm32f1::can_peripheral_manager` has its original constructor
deprecated and a new constructor accepting a `hal::steady_clock` and
timeout time. If exiting initialization takes longer than the time out
time, a `hal::timed_out` exception will be thrown.

In order to prevent applications unassociated from the can application
to continue to work as usual, a try/catch is put around the construction
of the `hal::stm32f1::can_peripheral_manager` object. If construction
fails, then none of the can interface optionals will be set and the
program will run without them. If the can application is executed and
the optionals are nullopt, then access will throw and will terminate
the application which is the intended behavior.