:rocket: IT WORKS!!
The enumeration process is now very fast. The issue seems to come down
to missing packets/events and responding to the wrong events. Rather
overwriting the current command buffer, a circular buffer can be used to
manage new messages coming in while we attempt to handle the current.
I also believe that the usage of the ACT boolean was also to blame for
the issues and slowness. The ACT boolean was being changed while the
code was returning a response to the host. The problem is that the host
can be quick and respond with an answer around the time we are printing
to the serial port. At the end of the function we set act to false which
means we miss a SETUP packet. The host then waits for us to respond
then it send a new message. All of this confuses the host and it
sometimes attributes this to having "multiple configurations" and
will show up as "Composite Device" on mac using lsusb.