Use usb_endpoint_info for descriptors
NOTE: fixed the issue with data above endpoint size being ignored. This
was due to the endpoint length being incorrect. The previous
descriptor's length was much larger than 16 bytes and when the endpoint
was sent data longer than its length, it ignored these. Now with the
correct length, we get all of the data.