:sparkles: stm32f1::power_on() now throws (#104)
When power_on is called and the peripheral is already powered on, this
is considered a violation of the single use principle per each
peripheral, which means it will throw. This will detect more than one
driver attempting to control a single peripheral. Driver should power of
their peripherals if they are no longer using them in their destructor.

Resolves #103