I think I've seen all that I need
Coroutines functions pretty fit the size you'd expect for the operations
it conducts. We can make exceptions even faster to support the fact that
all coros re-entrant. We can compress the exception metadata even
further to ensure that the size does not grow for coroutines. I don't
think I'm concerned anymore about coroutines. I want to go forward
and use them in libhal for v5 for v6. We shall see. I'm excited to see
the kind of code we can write with this. One thing that I need to
add is a state and `u64` data field to async_context which communicates
what state the task is in such as:

```C++
enum class async_state : i32 {
    ready = 0,
    time_blocked = 1,
    io_blocked = 2,
    sync_blocked = 3,
    message_blocked = 4,
};
```

That way we can communicate to a scheduler what kind of blocking we
are currently blocked for. As an example `time_blocked` represents the
task needing to wait a time duration before it can continue.