:bug: (patch) Fix stm32f1 pwm driver move constructors (#148)
Previously, if a driver is an lvalue on the stack and its moved, the
moved from lvalue will destruct and disable the PWM while its at it.
This change updates the pwm drivers to have a move constructor that
sets the data pointers to nullptr so the dtor can detect a moved from
object.